@@ -1,16 +1,15 @@
 name: esp_rainmaker_association
 description: Native dart implementation of ESP Rainmaker provisioning and user mapping.
-version: 0.0.2
+version: 1.0.0
 homepage: https://github.com/chmoore889/esp_rainmaker_association
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  http: ^0.12.2
-  protobuf: ^1.1.0
+  http: ^0.13.3
+  protobuf: ^2.0.0
 
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.14.4
-  uuid: ^2.2.2
+  pedantic: ^1.11.1
+  uuid: ^3.0.4
