@@ -2,7 +2,7 @@
 //  Generated code. Do not modify.
 //  source: sec0.proto
 //
-// @dart = 2.3
+// @dart = 2.12
 // ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields
 
 import 'dart:core' as $core;
@@ -42,8 +42,8 @@ class S0SessionCmd extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   S0SessionCmd copyWith(void Function(S0SessionCmd) updates) =>
-      super.copyWith((message) =>
-          updates(message as S0SessionCmd)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as S0SessionCmd))
+          as S0SessionCmd; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static S0SessionCmd create() => S0SessionCmd._();
@@ -53,7 +53,7 @@ class S0SessionCmd extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static S0SessionCmd getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<S0SessionCmd>(create);
-  static S0SessionCmd _defaultInstance;
+  static S0SessionCmd? _defaultInstance;
 }
 
 class S0SessionResp extends $pb.GeneratedMessage {
@@ -78,7 +78,15 @@ class S0SessionResp extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   S0SessionResp._() : super();
-  factory S0SessionResp() => create();
+  factory S0SessionResp({
+    $0.Status? status,
+  }) {
+    final _result = create();
+    if (status != null) {
+      _result.status = status;
+    }
+    return _result;
+  }
   factory S0SessionResp.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -93,8 +101,8 @@ class S0SessionResp extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   S0SessionResp copyWith(void Function(S0SessionResp) updates) =>
-      super.copyWith((message) =>
-          updates(message as S0SessionResp)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as S0SessionResp))
+          as S0SessionResp; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static S0SessionResp create() => S0SessionResp._();
@@ -104,7 +112,7 @@ class S0SessionResp extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static S0SessionResp getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<S0SessionResp>(create);
-  static S0SessionResp _defaultInstance;
+  static S0SessionResp? _defaultInstance;
 
   @$pb.TagNumber(1)
   $0.Status get status => $_getN(0);
@@ -151,7 +159,23 @@ class Sec0Payload extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   Sec0Payload._() : super();
-  factory Sec0Payload() => create();
+  factory Sec0Payload({
+    Sec0MsgType? msg,
+    S0SessionCmd? sc,
+    S0SessionResp? sr,
+  }) {
+    final _result = create();
+    if (msg != null) {
+      _result.msg = msg;
+    }
+    if (sc != null) {
+      _result.sc = sc;
+    }
+    if (sr != null) {
+      _result.sr = sr;
+    }
+    return _result;
+  }
   factory Sec0Payload.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -166,8 +190,8 @@ class Sec0Payload extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   Sec0Payload copyWith(void Function(Sec0Payload) updates) =>
-      super.copyWith((message) =>
-          updates(message as Sec0Payload)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as Sec0Payload))
+          as Sec0Payload; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static Sec0Payload create() => Sec0Payload._();
@@ -176,10 +200,10 @@ class Sec0Payload extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static Sec0Payload getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<Sec0Payload>(create);
-  static Sec0Payload _defaultInstance;
+  static Sec0Payload? _defaultInstance;
 
   Sec0Payload_Payload whichPayload() =>
-      _Sec0Payload_PayloadByTag[$_whichOneof(0)];
+      _Sec0Payload_PayloadByTag[$_whichOneof(0)]!;
   void clearPayload() => clearField($_whichOneof(0));
 
   @$pb.TagNumber(1)
