@@ -2,7 +2,7 @@
 //  Generated code. Do not modify.
 //  source: esp_rmaker_user_mapping.proto
 //
-// @dart = 2.3
+// @dart = 2.12
 // ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields
 
 // ignore_for_file: UNDEFINED_SHOWN_NAME
@@ -34,7 +34,7 @@ class RMakerConfigStatus extends $pb.ProtobufEnum {
 
   static final $core.Map<$core.int, RMakerConfigStatus> _byValue =
       $pb.ProtobufEnum.initByValue(values);
-  static RMakerConfigStatus valueOf($core.int value) => _byValue[value];
+  static RMakerConfigStatus? valueOf($core.int value) => _byValue[value];
 
   const RMakerConfigStatus._($core.int v, $core.String n) : super(v, n);
 }
@@ -60,7 +60,7 @@ class RMakerConfigMsgType extends $pb.ProtobufEnum {
 
   static final $core.Map<$core.int, RMakerConfigMsgType> _byValue =
       $pb.ProtobufEnum.initByValue(values);
-  static RMakerConfigMsgType valueOf($core.int value) => _byValue[value];
+  static RMakerConfigMsgType? valueOf($core.int value) => _byValue[value];
 
   const RMakerConfigMsgType._($core.int v, $core.String n) : super(v, n);
 }
