@@ -2,7 +2,7 @@
 //  Generated code. Do not modify.
 //  source: sec1.proto
 //
-// @dart = 2.3
+// @dart = 2.12
 // ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields
 
 // ignore_for_file: UNDEFINED_SHOWN_NAME
@@ -40,7 +40,7 @@ class Sec1MsgType extends $pb.ProtobufEnum {
 
   static final $core.Map<$core.int, Sec1MsgType> _byValue =
       $pb.ProtobufEnum.initByValue(values);
-  static Sec1MsgType valueOf($core.int value) => _byValue[value];
+  static Sec1MsgType? valueOf($core.int value) => _byValue[value];
 
   const Sec1MsgType._($core.int v, $core.String n) : super(v, n);
 }
