@@ -2,9 +2,14 @@
 //  Generated code. Do not modify.
 //  source: sec1.proto
 //
-// @dart = 2.3
-// ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields
+// @dart = 2.12
+// ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields,deprecated_member_use_from_same_package
 
+import 'dart:core' as $core;
+import 'dart:convert' as $convert;
+import 'dart:typed_data' as $typed_data;
+
+@$core.Deprecated('Use sec1MsgTypeDescriptor instead')
 const Sec1MsgType$json = const {
   '1': 'Sec1MsgType',
   '2': const [
@@ -15,6 +20,10 @@ const Sec1MsgType$json = const {
   ],
 };
 
+/// Descriptor for `Sec1MsgType`. Decode as a `google.protobuf.EnumDescriptorProto`.
+final $typed_data.Uint8List sec1MsgTypeDescriptor = $convert.base64Decode(
+    'CgtTZWMxTXNnVHlwZRIUChBTZXNzaW9uX0NvbW1hbmQwEAASFQoRU2Vzc2lvbl9SZXNwb25zZTAQARIUChBTZXNzaW9uX0NvbW1hbmQxEAISFQoRU2Vzc2lvbl9SZXNwb25zZTEQAw==');
+@$core.Deprecated('Use sessionCmd1Descriptor instead')
 const SessionCmd1$json = const {
   '1': 'SessionCmd1',
   '2': const [
@@ -28,6 +37,10 @@ const SessionCmd1$json = const {
   ],
 };
 
+/// Descriptor for `SessionCmd1`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List sessionCmd1Descriptor = $convert.base64Decode(
+    'CgtTZXNzaW9uQ21kMRIsChJjbGllbnRfdmVyaWZ5X2RhdGEYAiABKAxSEGNsaWVudFZlcmlmeURhdGE=');
+@$core.Deprecated('Use sessionResp1Descriptor instead')
 const SessionResp1$json = const {
   '1': 'SessionResp1',
   '2': const [
@@ -49,6 +62,10 @@ const SessionResp1$json = const {
   ],
 };
 
+/// Descriptor for `SessionResp1`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List sessionResp1Descriptor = $convert.base64Decode(
+    'CgxTZXNzaW9uUmVzcDESKQoGc3RhdHVzGAEgASgOMhEuZXNwcmVzc2lmLlN0YXR1c1IGc3RhdHVzEiwKEmRldmljZV92ZXJpZnlfZGF0YRgDIAEoDFIQZGV2aWNlVmVyaWZ5RGF0YQ==');
+@$core.Deprecated('Use sessionCmd0Descriptor instead')
 const SessionCmd0$json = const {
   '1': 'SessionCmd0',
   '2': const [
@@ -56,6 +73,10 @@ const SessionCmd0$json = const {
   ],
 };
 
+/// Descriptor for `SessionCmd0`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List sessionCmd0Descriptor = $convert.base64Decode(
+    'CgtTZXNzaW9uQ21kMBIjCg1jbGllbnRfcHVia2V5GAEgASgMUgxjbGllbnRQdWJrZXk=');
+@$core.Deprecated('Use sessionResp0Descriptor instead')
 const SessionResp0$json = const {
   '1': 'SessionResp0',
   '2': const [
@@ -72,6 +93,10 @@ const SessionResp0$json = const {
   ],
 };
 
+/// Descriptor for `SessionResp0`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List sessionResp0Descriptor = $convert.base64Decode(
+    'CgxTZXNzaW9uUmVzcDASKQoGc3RhdHVzGAEgASgOMhEuZXNwcmVzc2lmLlN0YXR1c1IGc3RhdHVzEiMKDWRldmljZV9wdWJrZXkYAiABKAxSDGRldmljZVB1YmtleRIjCg1kZXZpY2VfcmFuZG9tGAMgASgMUgxkZXZpY2VSYW5kb20=');
+@$core.Deprecated('Use sec1PayloadDescriptor instead')
 const Sec1Payload$json = const {
   '1': 'Sec1Payload',
   '2': const [
@@ -124,3 +149,7 @@ const Sec1Payload$json = const {
     const {'1': 'payload'},
   ],
 };
+
+/// Descriptor for `Sec1Payload`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List sec1PayloadDescriptor = $convert.base64Decode(
+    'CgtTZWMxUGF5bG9hZBIoCgNtc2cYASABKA4yFi5lc3ByZXNzaWYuU2VjMU1zZ1R5cGVSA21zZxIqCgNzYzAYFCABKAsyFi5lc3ByZXNzaWYuU2Vzc2lvbkNtZDBIAFIDc2MwEisKA3NyMBgVIAEoCzIXLmVzcHJlc3NpZi5TZXNzaW9uUmVzcDBIAFIDc3IwEioKA3NjMRgWIAEoCzIWLmVzcHJlc3NpZi5TZXNzaW9uQ21kMUgAUgNzYzESKwoDc3IxGBcgASgLMhcuZXNwcmVzc2lmLlNlc3Npb25SZXNwMUgAUgNzcjFCCQoHcGF5bG9hZA==');
