@@ -2,9 +2,14 @@
 //  Generated code. Do not modify.
 //  source: session.proto
 //
-// @dart = 2.3
-// ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields
+// @dart = 2.12
+// ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields,deprecated_member_use_from_same_package
 
+import 'dart:core' as $core;
+import 'dart:convert' as $convert;
+import 'dart:typed_data' as $typed_data;
+
+@$core.Deprecated('Use secSchemeVersionDescriptor instead')
 const SecSchemeVersion$json = const {
   '1': 'SecSchemeVersion',
   '2': const [
@@ -13,6 +18,10 @@ const SecSchemeVersion$json = const {
   ],
 };
 
+/// Descriptor for `SecSchemeVersion`. Decode as a `google.protobuf.EnumDescriptorProto`.
+final $typed_data.Uint8List secSchemeVersionDescriptor = $convert.base64Decode(
+    'ChBTZWNTY2hlbWVWZXJzaW9uEg4KClNlY1NjaGVtZTAQABIOCgpTZWNTY2hlbWUxEAE=');
+@$core.Deprecated('Use sessionDataDescriptor instead')
 const SessionData$json = const {
   '1': 'SessionData',
   '2': const [
@@ -47,3 +56,7 @@ const SessionData$json = const {
     const {'1': 'proto'},
   ],
 };
+
+/// Descriptor for `SessionData`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List sessionDataDescriptor = $convert.base64Decode(
+    'CgtTZXNzaW9uRGF0YRI0CgdzZWNfdmVyGAIgASgOMhsuZXNwcmVzc2lmLlNlY1NjaGVtZVZlcnNpb25SBnNlY1ZlchIsCgRzZWMwGAogASgLMhYuZXNwcmVzc2lmLlNlYzBQYXlsb2FkSABSBHNlYzASLAoEc2VjMRgLIAEoCzIWLmVzcHJlc3NpZi5TZWMxUGF5bG9hZEgAUgRzZWMxQgcKBXByb3Rv');
