@@ -4,17 +4,16 @@ import 'package:http/http.dart';
 import 'package:esp_rainmaker_association/src/constants.dart';
 
 class SoftAPTransport {
-  final String baseUrl = ESPConstants.DEFAULT_WIFI_BASE_URL;
-  Client client;
+  static const String baseUrl = ESPConstants.DEFAULT_WIFI_BASE_URL;
+  final Client _client;
+
+  SoftAPTransport() : _client = Client();
 
   Future<Uint8List> sendConfigData(String path, Uint8List data) async {
-    client ??= Client();
-
     Uint8List responseBytes;
     try {
-      final url = 'http://$baseUrl/$path';
-      final resp = await client.post(
-        url,
+      final resp = await _client.post(
+        Uri.http(baseUrl, '/$path'),
         headers: {
           'Accept': 'text/plain',
           'Content-type': 'application/x-www-form-urlencoded',
@@ -24,6 +23,8 @@ class SoftAPTransport {
 
       if (resp.statusCode == 200) {
         responseBytes = resp.bodyBytes;
+      } else {
+        throw 'Network Error';
       }
     } catch (e) {
       print('HTTP: $e');
