@@ -1,13 +1,19 @@
 /// Represents an access point detected by the rainmaker device.
 class WiFiAccessPoint {
   /// SSID of the access point.
-  String wifiName;
+  final String wifiName;
 
   /// RSSI of the access point.
   int rssi;
 
   /// Security type of the access point.
-  int security;
+  final int security;
+
+  WiFiAccessPoint({
+    required this.wifiName,
+    required this.rssi,
+    required this.security,
+  });
 
   @override
   String toString() {
